requirements.txt
asyncpg==0.30.0
httpx==0.27.2
python-dotenv==1.0.1
